Move logp code into pymc3.distributions.logp and graph functions into pymc3.aesaraf
