+++ pymc3/model.py
@@ -35,10 +35,10 @@ from pandas import Series
 
 import pymc3 as pm
 
-from pymc3.aesaraf import generator, gradient, hessian, inputvars
+from pymc3.aesaraf import change_rv_size, generator, gradient, hessian, inputvars
 from pymc3.blocking import DictToArrayBijection, RaveledVars
 from pymc3.data import GenTensorVariable, Minibatch
-from pymc3.distributions import change_rv_size, logp_transform, logpt, logpt_sum
+from pymc3.distributions import logp_transform, logpt, logpt_sum
 from pymc3.exceptions import ImputationWarning
 from pymc3.math import flatten_list
 from pymc3.util import UNSET, WithMemoization, get_var_name
